python/konverter/requirements.txt
fastapi[standard]==0.116.2
uvicorn[standard]==0.36.0
python-multipart==0.0.20
Pillow==11.3.0
PyMuPDF==1.26.4
pillow-avif-plugin==1.5.2
pillow-heif==1.1.0
jinja2==3.1.6